refactor(project): rename packaged.
