+++ src/main/java/br/com/fiapbook/application/usecase/user/GetAllUsersUseCase.java
@@ -1,7 +1,7 @@
-package br.com.fiapbook.application.user.usecase;
+package br.com.fiapbook.application.usecase.user;
 
-import br.com.fiapbook.domain.user.entity.User;
-import br.com.fiapbook.domain.user.service.UserService;
+import br.com.fiapbook.domain.entity.user.User;
+import br.com.fiapbook.domain.service.user.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
