+++ src/main/java/br/com/fiapbook/domain/repository/user/UserRepository.java
@@ -1,6 +1,6 @@
-package br.com.fiapbook.domain.user.repository;
+package br.com.fiapbook.domain.repository.user;
 
-import br.com.fiapbook.domain.user.entity.User;
+import br.com.fiapbook.domain.entity.user.User;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
