+++ src/test/java/br/com/fiapbook/application/usecase/user/GetAllUsersUseCaseTest.java
@@ -4,9 +4,8 @@ import static br.com.fiapbook.shared.testData.user.UserTestData.getUserWithId;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
-import br.com.fiapbook.application.user.usecase.GetAllUsersUseCase;
-import br.com.fiapbook.domain.user.entity.User;
-import br.com.fiapbook.domain.user.service.UserService;
+import br.com.fiapbook.domain.entity.user.User;
+import br.com.fiapbook.domain.service.user.UserService;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
